requirements.txt
aiohttp==3.9.5
fastapi==0.111.0
pdfplumber==0.11.0
python-docx==1.1.2
uvicorn==0.29.0
alembic==1.13.1
inflection==0.5.1
psycopg2

google-api-python-client==2.127.0
google-auth-oauthlib==1.2.0